% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{add_taxonomy_columns}
\alias{add_taxonomy_columns}
\title{Add taxonomy annotations to blast tabular output}
\usage{
add_taxonomy_columns(
  blast_output_tbl,
  subject_acc_colname = "subject_acc_ver",
  ncbi_acc_key = NULL,
  taxonomy_level = "kingdom",
  map_superkindom = TRUE
)
}
\arguments{
\item{blast_output_tbl}{a tbl of blast output format 7 (\code{-outfmt 7})}

\item{subject_acc_colname}{a string (default : "subject_acc_ver")denoting column name of subject_acc_colname.}

\item{ncbi_acc_key}{user specific ENTREZ api key. Get one via \code{taxize::use_entrez()}}

\item{taxonomy_level}{a string indicating level of taxonomy to be mapped. Can be one of the following
\enumerate{
\item superkingdom
\item kingdom
\item phylum
\item subphylum
\item class
\item subclass
\item infraclass
\item cohort
\item order
\item suborder
\item infraorder
\item superfamily
\item family
\item subfamily
\item genus
\item species
\item tribe
\item no rank
}}

\item{map_superkindom}{logical (default TRUE). map superkingdom if kingdom not found. Valid only when taxonomy_level == "kingdom".}
}
\value{

}
\description{
Given the blast output (\code{-outfmt 7}) in a tabular format add taxonomic annotations to each subject hit.
}
