% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporter.R
\name{cat_green_tick}
\alias{cat_green_tick}
\title{Print green tick on terminal. Wrapper function on  cli::cat_bullet}
\usage{
cat_green_tick(...)
}
\arguments{
\item{...}{Other arguments to be passed to \code{cli::cat_bullet}}
}
\value{

}
\description{
Print green tick on terminal. Wrapper function on  cli::cat_bullet
}
