% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent_tree.R
\name{coalescent_tree}
\alias{coalescent_tree}
\title{Construct Phylogenetic Tree from Processed Sequence Files with Coalescent Method}
\usage{
coalescent_tree(
  seq.files,
  data_dir = NULL,
  seq.type = "DNA",
  tree_method = "ML",
  show_tree = TRUE
)
}
\arguments{
\item{seq.files}{A character vector of sequence file names (without extension)
to be used for constructing the phylogenetic tree.}

\item{data_dir}{The directory containing the sequence files. This is a required parameter.}

\item{seq.type}{The type of sequence to be processed, either "DNA" or "protein".
Default is "DNA".}

\item{tree_method}{The method for constructing the phylogenetic tree. Options include:
"ML" (Maximum Likelihood), "NJ" (Neighbor Joining), "UPGMA", "MP" (Maximum Parsimony),
and "BI" (Bayesian Inference). Default is "ML".}

\item{show_tree}{Logical, if TRUE the constructed supertree is displayed. Default is TRUE.}
}
\value{
A phylogenetic tree object (of class \code{phylo}), representing the supertree constructed from
the individual trees built from the sequence files.

If \code{show_tree} is TRUE, the function plots the coalescence tree.
If FALSE, it returns the constructed phylogenetic tree as a "multiPhylo" object.
}
\description{
It constructs a phylogenetic tree from multiple sequence files with coalescent method,
either DNA or protein sequences, by applying various tree-building methods
such as Maximum Likelihood (ML), Neighbor Joining (NJ), UPGMA, Maximum Parsimony (MP),
and Bayesian Inference (BI). It processes sequence files, constructs individual trees
based on the specified method, and combines them into a supertree.
The tree can be displayed if specified.
}
\details{
It performs the following steps:
\enumerate{
\item Retrieves sequence data from the provided files and processes them based on the specified sequence type.
\item Builds an individual phylogenetic tree for each sequence file using the chosen tree-building method.
\item Combines the individual trees into a supertree.
\item Optionally, displays the supertree if the \code{show_tree} parameter is TRUE.
\item Returns the supertree object.
}
}
\examples{
dna <- c("K01939", "K03644", "K00797", "K00927", "K00088", "K02257", "K00164",
         "K00820", "K06158", "K00008")
data_dir <- system.file("extdata", "sequences", package = "phyloR")
tree1 <- coalescent_tree(seq.file = dna,
                         seq.type = "protein",
                         data_dir = data_dir,
                         tree_method = "NJ",
                         show_tree = TRUE)
tree2 <- coalescent_tree(seq.file = dna,
                         seq.type = "DNA",
                         data_dir = data_dir,
                         tree_method = "UPGMA")

}
