% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{filter_blast_hits}
\alias{filter_blast_hits}
\title{Filter blast hits}
\usage{
filter_blast_hits(
  blast_tbl,
  evalue = NULL,
  bit_score = NULL,
  query_cov = NULL,
  identity = NULL,
  query_length = NULL
)
}
\arguments{
\item{blast_tbl}{an object of class tbl from blast output format 7. Column names must be identical to the given below
\enumerate{
\item query_acc_ver
\item subject_acc_ver
\item identity
\item alignment_length
\item mismatches
\item gap_opens
\item q_start
\item q_end
\item s_start
\item s_end
\item evalue
\item bit_score
\item positives
}

It can be assigned using the function \code{get_blast_outformat_7_colnames()}}

\item{evalue}{a numeric value; default NULL. Only hits having evalue <= than given will be retained.}

\item{bit_score}{a numeric value; default NULL. Only hits having bit_score >= than given will be retained.}

\item{query_cov}{a numeric value between 0 and 100 ; default NULL. Only hits having query_cov >= than given will be retained. query_cov will be calculated from the column \code{q_start} \code{q_end} and param \code{query_length}}

\item{identity}{a numeric value between 0 and 100;default NULL. Only hits having identity >= than given will be retained.}

\item{query_length}{a numeric value indicating length of query sequence.}
}
\value{
a tbl of filtered blast hits.
}
\description{
Filter blast hits
}
\examples{
\dontrun{
 f <- system.file("extdata","example_blast_output_01.txt", package = "phyloR")
 d <- readr::read_delim(f, delim ="\t" , col_names = F)
 colnames(d) <- phyloR::get_blast_outformat_7_colnames()
 filtered <- filter_blast_hits(d , query_cov = 90 , identity = 60, query_length = 249, evalue = 1e-6)
}


}
