% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{genbank2uid_tbl}
\alias{genbank2uid_tbl}
\title{A Wrapper function around \link{taxize::genbank2uid}.}
\usage{
genbank2uid_tbl(x, ...)
}
\arguments{
\item{x}{vector of ncbi gene bank accession alphanumeric string, or a gi numeric string \code{(x)}.}

\item{...}{other parameters to be passed to \code{taxize::genbank2uid}.}
}
\value{
a tbl with colnames x, taxid, class, match, multiple_matches, pattern_match, uri, name
}
\description{
Given a ncbi gene bank accession alphanumeric string or a gi numeric string \code{(x)}, it returns a tibble contaning ncbi taxonomy id, name and other related columns.
}
\examples{
\dontrun{
x <- c("XP_022900619.1", "XP_022900618.1", "XP_018333511.1", "XP_018573075.1")
genbank2uid_tbl(x = x)
}
}
