% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg_sequences.R
\name{get_kegg_sequences}
\alias{get_kegg_sequences}
\title{Retrieve Gene Sequences from KEGG Database}
\usage{
get_kegg_sequences(
  gene_ids,
  id.type = "KO_id",
  seq.type = "protein",
  species.list = NULL,
  species.type = "scientificname"
)
}
\arguments{
\item{gene_ids}{A character vector of gene identifiers (e.g., KO IDs, NCBI gene IDs, or KEGG gene IDs).}

\item{id.type}{A character string indicating the type of the provided gene IDs.
Options are "KO_id", "ncbi_id", or "kegg_id" (default is "KO_id").}

\item{seq.type}{A character string specifying the type of sequence to retrieve.
Options are "DNA" or "protein" (default is "protein").}

\item{species.list}{A character vector of species names to filter results by species.Normally not NULL only if gene_ids is "KO_id".
If NULL, no filtering is applied (default is NULL).}

\item{species.type}{A character string specifying the type of species identifier.
Options are "scientificname", "taxonomic_id", or "abbspname" (default is "scientificname").}
}
\value{
A list of gene sequences (either DNA or protein) corresponding to the input gene IDs.
If \code{seq.type} is "DNA", it returns nucleotide sequences, otherwise protein sequences.
}
\description{
It retrieves DNA or protein sequences for a given list of gene identifiers
from the KEGG database based on the provided gene IDs, species, and sequence type.
It can handle different types of gene identifiers and allows filtering by species.
}
\examples{
# Example 1: Retrieve protein sequences for a list of KO IDs
seq_set1 <- get_kegg_sequences(gene_ids = "K00161",
                               id.type = "KO_id",
                               seq.type = "DNA",
                               species.list = c("hsa", "ptr"),
                               species.type = "abbspname")

# Example 2: Retrieve DNA sequences for a list of NCBI gene IDs
seq_set2 <- get_kegg_sequences(gene_ids = c("5160", "465525"),
                               id.type = "ncbi_id")

# Example 3: Retrieve protein sequences for a list of KEGG IDS
seq_set3 <- get_kegg_sequences(gene_ids = c("hsa:5160", "ptr:465525"),
                               id.type = "kegg_id",
                               seq.type = "DNA")


}
