% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthologs.R
\name{get_orthologs}
\alias{get_orthologs}
\title{Get Orthologs for a Gene}
\usage{
get_orthologs(
  gene_id,
  id.type = "KO_id",
  species.list = NULL,
  species.type = "scientificname"
)
}
\arguments{
\item{gene_id}{A character string specifying the gene identifier (e.g., KO_id, ncbi_id, or kegg_id).}

\item{id.type}{A character string specifying the type of the gene identifier. Options are "KO_id", "ncbi_id", and "kegg_id". Default is "KO_id".}

\item{species.list}{A character vector or string specifying the species or species identifiers to filter the results. Default is NULL.}

\item{species.type}{A character string specifying the type of species identifier. Options are "scientificname", "taxonomic_id", and "abbspname". Default is "scientificname".}
}
\value{
A data frame containing the orthologous genes with columns for Entrez ID, Gene Symbol, and Species. If species is specified, only the orthologs for that species will be returned.
}
\description{
It retrieves orthologous genes for a given gene ID. The function supports different gene ID types (KO_id, ncbi_id, kegg_id) and allows filtering by species.
}
\examples{
taxinfo1 <- get_orthologs(gene_id = "K00161",
                          id.type = "KO_id",
                          species.list = "Homo sapiens",
                          species.type = "scientificname")
taxinfo2 <- get_orthologs(gene_id = "5160",
                          id.type = "ncbi_id",
                          species.list = "hsa",
                          species.type = "abbspname")

taxinfo3 <- get_orthologs(gene_id = "hsa:5160",
                          id.type = "kegg_id",
                          species.list = "9606",
                          species.type = "taxonomic_id")

}
