% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{get_query_cov}
\alias{get_query_cov}
\title{Calculate \% query coverage.}
\usage{
get_query_cov(qstart, qend, qlen)
}
\arguments{
\item{qstart}{A numeric denoting query start}

\item{qend}{A numeric denoting query end}

\item{qlen}{A numeric denoting query length}
}
\value{
A numeric
}
\description{
Given a sequence start, end and length coverage can be calculated.
}
\seealso{
\link{get_subj_cov}
}
