% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{get_subj_cov}
\alias{get_subj_cov}
\title{Calculate subject coverage.}
\usage{
get_subj_cov(sstart, send, slen)
}
\arguments{
\item{sstart}{A numeric denoting subject start}

\item{send}{A numeric denoting subject end}

\item{slen}{A numeric denoting subject length}
}
\value{
A numeric
}
\description{
Calculate subject coverage.
}
