% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{get_taxon_rank}
\alias{get_taxon_rank}
\title{Get ncbi phylogeny level for a given ncbi taxonomy id.}
\usage{
get_taxon_rank(x, rank = "kingdom")
}
\arguments{
\item{x}{a vector of valid ncbi taxonomy id.}

\item{rank}{a string denoting required ncbi phylogenetic level. Default "kingdom". An input can be one of the followings
\enumerate{
\item superkingdom
\item kingdom
\item phylum
\item subphylum
\item class
\item subclass
\item infraclass
\item cohort
\item order
\item suborder
\item infraorder
\item superfamily
\item family
\item subfamily
\item genus
\item species
\item tribe
\item no rank
}}
}
\value{
a tbl with the below column.
\enumerate{
\item query_taxon
\item kingdom
\item kingdom_id
\item rank
}
}
\description{
Get ncbi phylogeny level for a given ncbi taxonomy id.
}
\examples{
\dontrun{
x <- c(166361, 166361, 224129, 217634)
get_taxon_rank(x)
}

}
