% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporter.R
\name{null_or_character}
\alias{null_or_character}
\title{Check value NULL or character}
\usage{
null_or_character(x)
}
\arguments{
\item{x}{a value to be checked for character or NULL}
}
\value{

}
\description{
Check value NULL or character
}
\keyword{internal}
