% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporter.R
\name{null_or_logical}
\alias{null_or_logical}
\title{Check value NULL or LOGICAL}
\usage{
null_or_logical(x)
}
\arguments{
\item{x}{a value to be checked for logical or NULL}
}
\value{

}
\description{
Check value NULL or LOGICAL
}
\keyword{internal}
