% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporter.R
\name{null_or_numeric}
\alias{null_or_numeric}
\title{Check value NULL or Numeric}
\usage{
null_or_numeric(x)
}
\arguments{
\item{x}{a value to be checked for numeric or NULL}
}
\description{
Check value NULL or Numeric
}
\keyword{internal}
