% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{remove_redundancy}
\alias{remove_redundancy}
\title{Remove redundancy from blast tabular output}
\usage{
remove_redundancy(
  blast_output_tbl,
  subject_acc_colname = "subject_acc_ver",
  alignment_length_colname = "alignment_length"
)
}
\arguments{
\item{blast_output_tbl}{an object of class tbl containing blast tabular output}

\item{subject_acc_colname}{a string denoting a column of subject hits in a \code{blast_output_tbl}. Default "subject_acc_ver"}

\item{alignment_length_colname}{a string denoting a column of alignment length in a \code{blast_output_tbl}. Default "alignment_length"}
}
\value{
a tbl
}
\description{
Given the output of blast tabular format in a tbl, it removes redundant hits by subject hit ids. For each redundant hit longest hit will be retuned.
}
