% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{remove_redundant_hits}
\alias{remove_redundant_hits}
\title{Remove redundant hits from blast tabular output}
\usage{
remove_redundant_hits(
  blast_output_tbl,
  subject_acc_colname = "subject_acc_ver",
  subject_start_colname = "s_start",
  subject_end_colname = "s_end",
  keep_length = FALSE
)
}
\arguments{
\item{blast_output_tbl}{an object of class tbl containing blast tabular output.}

\item{subject_acc_colname}{a string denoting a column of subject hits in a \code{blast_output_tbl}. Default "subject_acc_ver".}

\item{subject_start_colname}{a string denoting a column of subject start. Default "s_start".}

\item{subject_end_colname}{a string denoting a column of subject end. Default "s_end".}

\item{keep_length}{logical, default FALSE, indicates whether to keep column of subject length.}
}
\value{
a tbl
}
\description{
Given the output of blast tabular format in a tbl, it removes redundant hits by subject accession.
For each redundant hit longest hit will be kept.
}
\examples{
\dontrun{
 f <- system.file("extdata","blast_output_01.txt" ,package = "phyloR")
 d <- readr::read_delim(f, delim ="\t" , col_names = F , comment = "#")
 colnames(d) <- phyloR::get_blast_outformat_7_colnames()
 remove_redundant_hits(d)
}
}
