% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{subset_bstringset}
\alias{subset_bstringset}
\title{get records from BStringset}
\usage{
subset_bstringset(x, y, partial_match = TRUE)
}
\arguments{
\item{x}{character vector of query ids}

\item{y}{an object of class \code{BStringset} from which sequences to be filtered}

\item{partial_match}{logical (default TRUE).
\enumerate {
\item TRUE : return sequences if element of x matches anywhere in the headers of y.
\item FALSE : return sequences only of exact match occur between element of x and headers of y.
}}
}
\value{
an object of class BStringset
}
\description{
an object of class BStringset stores sequences and headers from FASTA or FASTQ files. See \link{Biostrings::readBStringSet()}. This function helps users to filter sequences by headers with full or partial match.
}
