% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{tips_by_clade}
\alias{tips_by_clade}
\title{Retrieve tree tips by clade (node) number from a phylogenetic tree.}
\usage{
tips_by_clade(x, y)
}
\arguments{
\item{x}{an object of a class phylo.}

\item{y}{a numerical vector denoting tree nodes.}
}
\value{
a tibble with two columns 1) tips 2) clade_id
}
\description{
You can subset tree tips by node / clade number from a given tree.
}
\examples{

}
