% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{add_taxonomy_columns}
\alias{add_taxonomy_columns}
\title{Add taxonomy annotations to blast tabular output}
\usage{
add_taxonomy_columns(
  tbl,
  ncbi_accession_colname = "ncbi_accession",
  ncbi_acc_key = NULL,
  taxonomy_level = "kingdom",
  map_superkindom = TRUE,
  batch_size = 20
)
}
\arguments{
\item{tbl}{an object of class tbl}

\item{ncbi_accession_colname}{a string (default : "ncbi_accession") denoting column name of ncbi accession.}

\item{ncbi_acc_key}{user specific ENTREZ api key. Get one via \code{taxize::use_entrez()}}

\item{taxonomy_level}{a string indicating level of taxonomy to be mapped. Can be one of the following
\enumerate{
\item superkingdom
\item kingdom
\item phylum
\item subphylum
\item class
\item subclass
\item infraclass
\item cohort
\item order
\item suborder
\item infraorder
\item superfamily
\item family
\item subfamily
\item genus
\item species
\item tribe
\item no rank
}}

\item{map_superkindom}{logical (default TRUE). map superkingdom if kingdom not found. Valid only when taxonomy_level == "kingdom".}

\item{batch_size}{The number of queries to submit at a time.}
}
\value{

}
\description{
Given a tbl with a column of valid ncbi accession,  the function can add taxonomic annotations for each row.
}
