% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_trim.R
\name{align_trim}
\alias{align_trim}
\title{Align and Trim DNA or Protein Sequences}
\usage{
align_trim(seq.file, seq.type = "DNA", method = "ClustalW", output_file = NULL)
}
\arguments{
\item{seq.file}{A character string specifying the path to the input sequence file (FASTA format).}

\item{seq.type}{A character string specifying the type of sequences. Options are "DNA" or "protein".
Default is "DNA".}

\item{method}{A character string specifying the alignment method to use. Options are "ClustalW", "Muscle", "Mafft" or "T-coffee".
Default is "ClustalW".}

\item{output_file}{A character string specifying the path to the output file where the trimmed alignment will be saved.
If NULL (default), the function returns the trimmed alignment as a sequence object.}
}
\value{
If \code{output_file} is NULL, returns the trimmed alignment as an object of class \code{DNAMultipleAlignment} or \code{AAMultipleAlignment},
depending on the sequence type. Otherwise, the trimmed alignment is written to the specified output file.
}
\description{
It reads DNA or protein sequences from a file, aligns them using a selected alignment method, trims the alignment,
and optionally saves the trimmed alignment to a file.
}
\examples{
# Align and trim DNA sequences
DNA_seq <- system.file("extdata", "DNA_sequences.fas", package = "phyloR")
align_trim(DNA_seq, seq.type = "DNA", method = "ClustalW")

# Align and trim protein sequences
protein_seq <- system.file("extdata", "protein_sequences.fas", package = "phyloR")
align_trim(protein_seq, seq.type = "protein", method = "ClustalW")

}
