% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescent_tree_fetch.R
\name{coalescent_tree_fetch}
\alias{coalescent_tree_fetch}
\title{Coalescent Tree Construction from Multiple Genes}
\usage{
coalescent_tree_fetch(
  gene_ids,
  id.type = "KO_id",
  species.list = NULL,
  species.type = "scientificname",
  seq.type = "DNA",
  align_method = "ClustalW",
  tree_method = "NJ",
  show_tree = TRUE
)
}
\arguments{
\item{gene_ids}{A character vector of gene IDs to be used for building the coalescent tree.}

\item{id.type}{The type of gene ID used(e.g., KO_id, ncbi_id, or kegg_id), default is "KO_id" (KEGG Orthology ID).}

\item{species.list}{A character vector specifying the species to be included in the analysis.
If NULL, all species available for each gene ID will be used.}

\item{species.type}{The type of species identifier to be used, Options are "scientificname", "taxonomic_id", and "abbspname".
default is "scientificname".}

\item{seq.type}{The type of sequence to be retrieved, default is "DNA".
Other options may include "protein".}

\item{align_method}{The alignment method to be used, default is "ClustalW".
Options are "ClustalW", "Muscle", "Mafft" or "T-coffee". Default is "ClustalW".}

\item{tree_method}{The method for constructing the phylogenetic tree. Options include:
"ML" (Maximum Likelihood), "NJ" (Neighbor Joining), "UPGMA", "MP" (Maximum Parsimony),
and "BI" (Bayesian Inference). Default is "ML".}

\item{show_tree}{Logical, if TRUE the constructed tree is displayed using a suitable tree plotting method.
Default is TRUE.}
}
\value{
A phylogenetic tree object (of class \code{phylo}), representing the coalescent tree constructed from
the aligned and trimmed sequences.
}
\description{
It constructs a coalescent phylogenetic tree by fetching gene sequences
from a given list of gene IDs, aligning and trimming them, and then building the tree
using a slected tree-building method. It retrieves orthologous gene sequences
for the slected species, processes them, and outputs a coalescent tree based on the selected
alignment and tree-building methods.
}
\details{
It performs the following steps:
\enumerate{
\item Retrieves orthologous gene sequences for the provided gene IDs and species.
\item Aligns and trims the sequences based on the slected alignment method.
\item Constructs a phylogenetic tree using the processed sequence data and the selected tree-building method.
\item Optionally, displays the tree if the \code{show_tree} parameter is set to TRUE.
\item Cleans up temporary files and directories after tree construction.
}
}
\examples{
tree <- coalescent_tree_fetch(gene_ids = c("K00820", "K00088", "K00927",
                                           "K06158", "K00008", "K00164",
                                           "K00797", "K01939", "K02257",
                                           "K03644"),
                              id.type = "KO_id",
                              species.list = c("ath", "gmx", "zma", "osa",
                                               "dme", "cel", "mmu", "rno",
                                               "hsa", "mcc", "ssc", "bta",
                                               "gga", "xla", "sce", "ece"),
                              species.type = "abbspname",
                              seq.type = "DNA",
                              tree_method = "NJ")

}
