% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trees.R
\name{compare_trees}
\alias{compare_trees}
\title{Compare Two Phylogenetic Trees}
\usage{
compare_trees(tree1, tree2)
}
\arguments{
\item{tree1}{A phylogenetic tree of class 'phylo', representing the first tree to compare.}

\item{tree2}{A phylogenetic tree of class 'phylo', representing the second tree to compare.}
}
\value{
A data frame containing the following similarity metrics:
\describe{
\item{RF_similarity}{Robinson-Foulds similarity, a measure of tree topology similarity (range: 0 to 1). A higher value indicates greater similarity in tree topology.}
\item{edge_similarity}{Edge similarity, representing the proportion of common edges between the two trees (range: 0 to 1). A value of 1 indicates identical edge structures.}
\item{entanglement_value}{Entanglement value, a metric quantifying the structural similarity of the trees, with higher values indicating greater similarity.}
}
}
\description{
It compares two phylogenetic trees by calculating their
similarity based on four different metrics: Robinson-Foulds (RF) similarity,
edge similarity, and entanglement value. The comparisons
are returned as a data frame containing the computed similarity values.
}
\details{
\itemize{
\item \strong{Robinson-Foulds Similarity}: This metric is based on the Robinson-Foulds (RF) distance, which compares the trees by evaluating their edge partitions. The RF similarity score is computed by normalizing the RF distance with respect to the maximum possible RF distance between the two trees.
\item \strong{Edge Similarity}: This metric measures the proportion of edges that are common between the two trees. It is calculated by comparing the edges of both trees and counting how many are identical.
\item \strong{Entanglement Value}: This metric calculates the structural similarity of the trees based on their dendrogram representations. A higher entanglement value indicates a greater similarity in tree structure.
}
}
\examples{
\dontshow{if (requireNamespace("dendextend", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Compare two phylogenetic trees
DNA_seq <- system.file("extdata", "DNA_seq.fas", package = "phyloR")
tree1 <- gene_tree(seq.file = DNA_seq,
                   seq.type = "DNA",
                   tree_method = "UPGMA")
species = c("ath", "gmx", "zma", "osa",
            "dme", "cel", "mmu", "rno",
            "hsa", "mcc", "ssc", "bta",
            "gga", "xla", "sce", "ece")
tree2 <- species_tree(species = species, species.type = "abbspname")
relation <- compare_trees(tree1, tree2)
\dontshow{\}) # examplesIf}
}
