% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trees.R
\name{compare_trees}
\alias{compare_trees}
\title{Compare Two Phylogenetic Trees}
\usage{
compare_trees(tree1, tree2)
}
\arguments{
\item{tree1}{A phylogenetic tree of class 'phylo'.}

\item{tree2}{A phylogenetic tree of class 'phylo'.}
}
\value{
A data frame containing the following similarity metrics:
\describe{
\item{RF_similarity}{Robinson-Foulds similarity, a measure of tree topology similarity (range: 0 to 1).}
\item{edge_similarity}{Edge similarity, representing the proportion of common edges between the trees (range: 0 to 1).}
\item{subtree_similarity}{Subtree similarity, based on the proportion of identical subtrees between the trees (range: 0 to 1).}
}
}
\description{
This function compares two phylogenetic trees by calculating their
similarity based on three different metrics: Robinson-Foulds (RF) similarity,
edge similarity, and subtree similarity. The comparisons are returned as a data frame.
}
\details{
\itemize{
\item \strong{Robinson-Foulds Similarity} is calculated by measuring the RF distance and converting it into a similarity score. The RF distance compares the trees based on their edge partitioning.
\item \strong{Edge Similarity} is determined by counting the common edges between the two trees and calculating the proportion of matching edges.
\item \strong{Subtree Similarity} is determined by comparing the subtrees rooted at internal nodes, counting how many of them are identical between the two trees.
}
}
\examples{
# Compare two phylogenetic trees
tree1 <- ape::rtree(100)
tree2 <- phangorn::rSPR(tree1, 3)
x <- compare_trees(tree1, tree2)

}
