% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_tree_fetch.R
\name{concat_tree_fetch}
\alias{concat_tree_fetch}
\title{Concatenate Phylogenetic Tree from Multiple Genes}
\usage{
concat_tree_fetch(
  gene_ids,
  id.type = "ko_id",
  species.list = NULL,
  species.type = "scientificname",
  seq.type = "DNA",
  align_method = "ClustalW",
  gapOpening = "default",
  gapExtension = "default",
  maxiters = "default",
  gap.end = 0.5,
  gap.mid = 0.9,
  tree_method = "NJ",
  model = NULL,
  show_tree = TRUE
)
}
\arguments{
\item{gene_ids}{A character vector of gene IDs to be used for building the concatenate tree.}

\item{id.type}{The type of gene ID used(e.g., ko_id, ncbi_id, or kegg_id), default is "ko_id" (KEGG Orthology ID).}

\item{species.list}{A character vector specifying the species to be included in the analysis.
If NULL, all species available for each gene ID will be used.}

\item{species.type}{The type of species identifier to be used, Options are "scientificname", "taxonomic_id", and "abbspname".
default is "scientificname".}

\item{seq.type}{The type of sequence to be retrieved, default is "DNA".
Other options may include "protein".
Options are "ClustalW" (default), "Muscle", or "ClustalOmega".}

\item{align_method}{The alignment method to be used, default is "ClustalW".
Options are "ClustalW", "Muscle" or "ClustalOmega". Default is "ClustalW".}

\item{gapOpening}{The penalty for opening a gap in the alignment.
Default is "default", which uses algorithm-specific default values.}

\item{gapExtension}{The penalty for extending an existing gap.
Default is "default", which uses algorithm-specific default values.}

\item{maxiters}{The maximum number of refinement iterations.
Default is "default", which uses algorithm-specific default values.}

\item{gap.end}{Fraction of gaps tolerated at the ends of the alignment (0-1). Default is 0.5.}

\item{gap.mid}{Fraction of gaps tolerated inside the alignment (0-1). Default is 0.9.}

\item{tree_method}{The method for constructing the phylogenetic tree. Options include:
"ML" (Maximum Likelihood), "NJ" (Neighbor Joining), "UPGMA", "MP" (Maximum Parsimony),
and "BI" (Bayesian Inference). Default is "ML".}

\item{show_tree}{Logical, if TRUE the constructed tree is displayed using a suitable tree plotting method.
Default is TRUE.}
}
\value{
A phylogenetic tree object of class \code{phylo}, representing the tree constructed from the concatenated sequences.
}
\description{
It fetches orthologous sequences for multiple gene IDs, performs alignment
and trimming, and constructs a phylogenetic tree based on the concatenated sequences.
It allows the selection of different sequence alignment and tree-building methods,
and supports both DNA and protein sequences.
}
\details{
It performs the following steps:
\enumerate{
\item Fetches orthologous sequences for each gene ID and species.
\item Aligns and trims the sequences using the slected alignment method.
\item Constructs a phylogenetic tree based on the concatenated sequences using the slected tree method.
\item Optionally displays the constructed tree.
\item Cleans up temporary files and restores the original working directory.
}
}
\examples{
\dontrun{
tree <- concat_tree_fetch(gene_ids = c("K00820", "K00088", "K00927",
                                       "K06158", "K00008","K00164",
                                       "K00797", "K01939", "K02257",
                                       "K03644"),
                          id.type = "ko_id",
                          species.list = c("ath", "gmx", "zma", "osa",
                                           "dme", "cel", "mmu", "rno",
                                           "hsa", "mcc", "ssc", "bta",
                                           "gga", "xla", "sce", "ece"),
                          species.type = "abbspname",
                          seq.type = "DNA",
                          tree_method = "UPGMA")
}


}
