% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{genbank2uid_tbl}
\alias{genbank2uid_tbl}
\title{Wrapper function around taxize::genbank2uid.}
\usage{
genbank2uid_tbl(x, ...)
}
\arguments{
\item{x}{vector of genBank accession alphanumeric string, or a gi numeric string \code{(x)}.}

\item{...}{other parameters to be passed to \code{taxize::genbank2uid}}
}
\value{
a tbl with colnames x, taxid, class, match, multiple_matches, pattern_match, uri, name
}
\description{
Given a genBank accession alphanumeric string, or a gi numeric string \code{(x)}, it returns tibble of taxid, name and other columns.
}
\examples{
\dontrun{
x <- c("XP_022900619.1", "XP_022900618.1", "XP_018333511.1", "XP_018573075.1")
genbank2uid_tbl(x = x)
}
}
