% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tree.R
\name{gene_tree}
\alias{gene_tree}
\title{Construct Phylogenetic Tree from Processed Single Gene Sequences File}
\usage{
gene_tree(
  seq.file,
  seq.type = "DNA",
  tree_method = "NJ",
  root = NULL,
  show_tree = TRUE
)
}
\arguments{
\item{seq.file}{A string specifying the path to the sequence file (FASTA format).}

\item{seq.type}{A string specifying the type of sequences in the file. Options are "DNA" or "protein". Default is "DNA".}

\item{tree_method}{A string specifying the tree construction method. Options include:
"NJ" (Neighbor-Joining), "MP" (Maximum-Parsimony), "ML" (Maximum-Likelihood), "UPGMA", "BI" (Bayesian-Inference).
Default is "NJ".}

\item{root}{Character string or NULL, specify outgroup for rooting. If NULL, automatic outgroup selection is performed}

\item{show_tree}{Logical, if TRUE the constructed phylogenetic tree is displayed. Default is TRUE.}
}
\value{
A phylogenetic tree object of class \code{phylo} representing the tree constructed using the selected method.
If \code{show_tree} is TRUE, the tree is plotted with bootstrap support values.
}
\description{
It generates a phylogenetic tree from sequence data using various tree construction methods,
including Neighbor-Joining (NJ), Maximum-Parsimony (MP), Maximum-Likelihood (ML), Unweighted Pair Group
Method with Arithmetic Mean (UPGMA), and Bayesian-Inference (BI). It supports both DNA and protein
sequences and allows for bootstrapping and tree visualization.
}
\details{
The function supports the following tree construction methods:
\itemize{
\item \strong{Neighbor-Joining (NJ)}: Constructs a tree using the NJ algorithm, with options for DNA or protein sequences.
\item \strong{Maximum-Parsimony (MP)}: Uses the MP algorithm for tree construction with parsimony optimization.
\item \strong{Maximum-Likelihood (ML)}: Implements ML-based tree construction, including model selection and bootstrapping.
\item \strong{Unweighted Pair Group Method with Arithmetic Mean (UPGMA)}: Generates a UPGMA tree based on pairwise distances.
\item \strong{Bayesian-Inference (BI)}: Constructs a Bayesian phylogenetic tree using BEAST2 and supports posterior probabilities.
}
}
\examples{
# Example of constructing trees using different methods from a DNA sequence file
DNA_seq <- system.file("extdata", "DNA_seq.fas", package = "phyloR")
tree1 <- gene_tree(seq.file = DNA_seq,
                   seq.type = "DNA",
                   tree_method = "UPGMA")

# Example of constructing trees using different methods from a protein sequence file
protein_seq <- system.file("extdata", "protein_seq.fas", package = "phyloR")
tree2 <- gene_tree(seq.file = protein_seq,
                   seq.type = "protein",
                   tree_method = "UPGMA")

}
