% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tree_fetch.R
\name{gene_tree_fetch}
\alias{gene_tree_fetch}
\title{Construct Phylogenetic Tree from Single Gene}
\usage{
gene_tree_fetch(
  gene_id,
  id.type = "ko_id",
  species.list,
  species.type = "scientificname",
  seq.type = "DNA",
  align_method = "ClustalW",
  tree_method = "NJ",
  show_tree = TRUE
)
}
\arguments{
\item{gene_id}{A string specifying the gene identifier (e.g., ko_id).}

\item{id.type}{The type of gene ID used(e.g., ko_id, ncbi_id, or kegg_id), default is "ko_id" (KEGG Orthology ID).}

\item{species.list}{A character vector specifying the species to be included in the analysis.
If NULL, all species available for each gene ID will be used.}

\item{species.type}{The type of species identifier to be used, Options are "scientificname", "taxonomic_id", and "abbspname".
default is "scientificname".}

\item{seq.type}{The type of sequence to be retrieved, default is "DNA".
Other options may include "protein".}

\item{align_method}{The alignment method to be used, default is "ClustalW".
Options are "ClustalW", "Muscle", "Mafft" or "T-coffee". Default is "ClustalW".}

\item{tree_method}{The method for constructing the phylogenetic tree. Options include:
"ML" (Maximum Likelihood), "NJ" (Neighbor Joining), "UPGMA", "MP" (Maximum Parsimony),
and "BI" (Bayesian Inference). Default is "ML".}

\item{show_tree}{Logical, if TRUE the constructed tree is displayed using a suitable tree plotting method.
Default is TRUE.}
}
\value{
A phylogenetic tree object (of class \code{phylo}), representing the coalescent tree constructed from
the aligned and trimmed sequences.
}
\description{
It retrieves orthologous gene for a list of species, performs sequence
alignment, and constructs a phylogenetic tree based on the concatenated sequences.
It supports different sequence types and alignment methods, with multiple tree construction options.
}
\details{
It performs the following steps:
\enumerate{
\item Retrieves orthologous gene sequences for the provided gene IDs and species.
\item Aligns and trims the sequences based on the specified alignment method.
\item Constructs a phylogenetic tree using the processed sequence data and the selected tree-building method.
\item Optionally, displays the tree if the \code{show_tree} parameter is set to TRUE.
\item Cleans up temporary files and directories after tree construction.
}
}
\examples{
# Example usage:
tree <- gene_tree_fetch(gene_id = "K00826",
                        species.list = c("hsa", "sce", "dme", "cel",
                                         "xla", "gga", "ssc", "rno",
                                         "mmu", "mcc", "gmx", "bta",
                                         "ece", "zma", "osa", "ath"),
                        species.type = "abbspname",
                        seq.type = "DNA",
                        tree_method = "NJ",
                        show_tree = TRUE)

}
