% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{get_blast_outformat_7_colnames}
\alias{get_blast_outformat_7_colnames}
\title{Get blast output format 7 column names}
\usage{
get_blast_outformat_7_colnames()
}
\value{
A character vector
}
\description{
Get blast output format 7 column names
}
\examples{
\dontrun{
get_blast_outformat_7_colnames()
}


}
