% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_result_parser.R
\name{get_cov}
\alias{get_cov}
\title{Calculate sequence coverage}
\usage{
get_cov(start, end, len)
}
\arguments{
\item{start}{A numeric denoting start}

\item{end}{A numeric denoting end}

\item{len}{A numeric denoting length}
}
\value{
A numeric
}
\description{
Calculate sequence coverage
}
\seealso{
get_query_cov

get_subj_cov
}
\keyword{internal}
