% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{get_taxon_rank}
\alias{get_taxon_rank}
\title{Title}
\usage{
get_taxon_rank(x, rank = "kingdom")
}
\arguments{
\item{x}{a vector of valid ncbi taxid}

\item{rank}{default "kingdom". Can be one of the "no rank", "superkingdom", "kingdom", "phylum", "subphylum", "class", "subclass", "infraclass", "cohort", "order", "suborder", "infraorder", "superfamily", "family", "subfamily", "genus", "species", "tribe"}
}
\value{
a tbl
}
\description{
Title
}
\examples{
\dontrun{
x <- c(166361, 166361, 224129, 217634)
get_taxon_rank(x)
}

}
