% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_tree.R
\name{species_tree}
\alias{species_tree}
\title{Construct Phylogenetic Tree from Species Data}
\usage{
species_tree(
  species,
  species.type = "scientificname",
  db = "ncbi",
  show_tree = TRUE
)
}
\arguments{
\item{species}{A character vector containing species names or taxonomic IDs. The input can be:
- scientific names (e.g., "Homo sapiens"),
- taxonomic IDs (numeric IDs),
- abbreviated species names (e.g., "hsa" for Homo sapiens).}

\item{species.type}{A string indicating the type of input species. Options are:
"scientificname" (default), "taxonomic_id", or "abbspname" (abbreviated species name).}

\item{db}{A string specifying the database to use for retrieving taxonomy information. Default is "ncbi".}

\item{show_tree}{Logical, if TRUE the constructed phylogenetic tree is plotted. Default is TRUE.}
}
\value{
A phylogenetic tree object of class \code{phylo} representing the tree constructed from the species data.
If \code{show_tree} is TRUE, the tree is displayed.
}
\description{
It generates a phylogenetic tree based on a list of species names or taxonomic IDs.
The tree is constructed using the taxonomy information retrieved from a specified database (e.g., NCBI).
It supports three input types for species: scientific names, taxonomic IDs, and abbreviated species names.
}
\details{
It retrieves species classification data from a database (e.g., NCBI) based on the species names or IDs
provided. It then constructs a phylogenetic tree using the taxonomy classification. The supported species input types
are scientific names, taxonomic IDs, and abbreviated species names.
}
\examples{
# Example 1:
species1 <- c("Homo sapiens", "Pan troglodytes", "Mus musculus",
              "Rattus norvegicus","Canis lupus familiaris", "Felis catus")
tree1 <- species_tree(species = species1, species.type = "scientificname")

# Example 2:
species2 <- c("9606", "9598", "10090", "9615", "9685", "10116")
tree2 <- species_tree(species = species2, species.type = "taxonomic_id")

# Example 3:
species3 = c("ath", "gmx", "zma", "osa",
             "dme", "cel", "mmu", "rno",
             "hsa", "mcc", "ssc", "bta",
             "gga", "xla", "sce", "ece")
tree3 <- species_tree(species = species3, species.type = "abbspname")

}
